refactor: try to make ivy compatible
It moves the project from rollup to the angular/cli library format.
This seems to break compatibility with ViewEngine builds and I could not get the tests to run.
As far as I could tell, it 'works' but I won't guarantee anything!

If anyone has the time to clean that up or anything, feel free.
I probably won't be able invest much time here.

BREAKING CHANGE: breaks ViewEngine compatibility